Split up mint auditor models into separate files, get rid of unnecessary traits
